Implemented lint_all_views command.
